Extract login form into AuthForm component
